requirements.txt
absl-py==1.0.0; python_version >= "3.6"
ase==3.22.1; python_version >= "3.6"
astunparse==1.6.3
cachetools==4.2.4; python_version >= "3.5" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3.7"
colorama==0.4.4; python_version >= "2.7" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and platform_system == "Windows"
cycler==0.11.0; python_version >= "3.7"
flatbuffers==2.0
fonttools==4.28.5; python_version >= "3.7"
future==0.18.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
gast==0.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
google-auth-oauthlib==0.4.6; python_version >= "3.6"
google-auth==2.3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-pasta==0.2.0
grpcio==1.43.0; python_version >= "3.6"
h5py==3.6.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
importlib-metadata==4.10.0; python_version < "3.10" and python_version >= "3.7"
joblib==1.1.0; python_version >= "3.7"
keras-preprocessing==1.1.2
keras==2.7.0
kiwisolver==1.3.2; python_version >= "3.7"
libclang==12.0.0
markdown==3.3.6; python_version >= "3.6"
matplotlib==3.5.1; python_version >= "3.7"
megnet==1.2.9
monty==2021.12.1; python_version >= "3.7"
mpmath==1.2.1; python_version >= "3.7"
networkx==2.6.3; python_version >= "3.7"
numpy==1.21.5
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
opt-einsum==3.3.0; python_version >= "3.5"
packaging==21.3; python_version >= "3.7"
palettable==3.3.0; python_version >= "3.7"
pandas==1.3.5; python_full_version >= "3.7.1"
pillow==8.4.0; python_version >= "3.7"
plotly==5.5.0; python_version >= "3.7"
protobuf==3.19.1; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pymatgen==2022.0.17; python_version >= "3.7"
pyparsing==3.0.6; python_version >= "3.7"
python-dateutil==2.8.2; python_full_version >= "3.7.1" and python_version >= "3.7"
pytz==2021.3; python_full_version >= "3.7.1" and python_version >= "3.7"
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.26.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.10" and python_version >= "3.7"
ruamel.yaml==0.17.17; python_version >= "3.7"
scikit-learn==1.0.1; python_version >= "3.7"
scipy==1.7.3; python_version >= "3.7" and python_version < "3.11"
setuptools-scm==6.3.2; python_version >= "3.7"
six==1.16.0; python_full_version >= "3.7.1" and python_version >= "3.7" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
sklearn==0.0
spglib==1.16.3; python_version >= "3.7"
sympy==1.9; python_version >= "3.7"
tabulate==0.8.9; python_version >= "3.7"
tenacity==8.0.1; python_version >= "3.7"
tensorboard-data-server==0.6.1; python_version >= "3.6"
tensorboard-plugin-wit==1.8.0; python_version >= "3.6"
tensorboard==2.7.0; python_version >= "3.6"
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.23.1; python_version >= "3.7" and python_version < "3.11"
termcolor==1.1.0
threadpoolctl==3.0.0; python_version >= "3.7"
tomli==2.0.0; python_version >= "3.7"
tqdm==4.62.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
typing-extensions==4.0.1; python_version >= "3.6"
uncertainties==3.1.6; python_version >= "3.7"
urllib3==1.26.7; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
werkzeug==2.0.2; python_version >= "3.6"
wrapt==1.13.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
zipp==3.6.0; python_version < "3.10" and python_version >= "3.7"
